+++ lib/bencher_json/src/system/mod.rs
@@ -4,4 +4,5 @@ pub mod config;
 pub mod endpoint;
 pub mod ping;
 pub mod restart;
+pub mod spec;
 pub mod version;