+++ lib/bencher_json/src/system/spec.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "schema")]
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+#[cfg_attr(feature = "schema", derive(JsonSchema))]
+pub struct JsonSpec(pub serde_json::Value);