+++ services/cli/src/bencher/sub/system/server/mod.rs
@@ -7,12 +7,14 @@ mod config;
 mod endpoint;
 mod ping;
 mod restart;
+mod spec;
 mod version;
 
 #[derive(Debug)]
 pub enum Server {
     Ping(ping::Ping),
     Version(version::Version),
+    Spec(spec::Spec),
     Endpoint(endpoint::Endpoint),
     Restart(restart::Restart),
     Config(config::Config),
@@ -26,6 +28,7 @@ impl TryFrom<CliServer> for Server {
         Ok(match admin {
             CliServer::Ping(ping) => Self::Ping(ping.try_into()?),
             CliServer::Version(version) => Self::Version(version.try_into()?),
+            CliServer::Spec(spec) => Self::Spec(spec.try_into()?),
             CliServer::Endpoint(endpoint) => Self::Endpoint(endpoint.try_into()?),
             CliServer::Restart(restart) => Self::Restart(restart.try_into()?),
             CliServer::Config(config) => Self::Config(config.try_into()?),
@@ -40,6 +43,7 @@ impl SubCmd for Server {
         match self {
             Self::Ping(ping) => ping.exec().await,
             Self::Version(version) => version.exec().await,
+            Self::Spec(spec) => spec.exec().await,
             Self::Endpoint(endpoint) => endpoint.exec().await,
             Self::Restart(restart) => restart.exec().await,
             Self::Config(config) => config.exec().await,